Completely seperated Interpreter and Environment. Also, modified every Command and Event to unpack everything they might need
Every Command and Event now unpacks whatever it might need from the
Environment during instantiation, and uses special functions to verify
that what it unpacked is valid. If it's not, it will add an error to
self.errors, and the GUI can then display any "Compile" errors in a nice
little message. That way if you have commands that use a robot, and your
robot is not plugged in, when you run the script it will warn you of
this fact. Same goes for cameras, settings/calibrations, or more
importantly- computer vision objects.