Added descriptive error messages when robot fails to connect
When the user attempts to connect to a uArm, the GUI will now display a descriptive error if the uArm fails to connect.

This fixes several issues. For example, in Linux, sometimes uArms will not connect because they require Root permission. However, this is not the case in Windows or Mac.

Furthermore, users might not realize that their robot is not plugged in, or plugged into the wrong port, so this will ensure they understand the issue.

If there is an error while connecting to the robot, the settings.txt will remove the robotID so the user will have to go to Devices to connect to their robot again.